backend/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
GitPython==3.1.32
docker==6.1.3
requests==2.31.0
pyyaml==6.0.1
